+++ Changes.md
@@ -1,5 +1,5 @@
 # master
-- TypeScript: type children of imported components as `React.ReactNode`.
+- TypeScript: type children of imported components as `React.ReactNode` instead of `JSX.Element`.
 
 # 4.2.0
 - Add basic support for inherited type definitions in polymorphic variants e.g. `type color = [red | blue]`. No support for conversion at the moment.